Minor code fixes

+++ src/main/java/de/blau/android/views/util/LRUMapTileCache.java
@@ -200,7 +200,7 @@ public class LRUMapTileCache {
 				// see if we can expand by 50%
 				if (maxCacheSize < (Runtime.getRuntime().maxMemory() - Runtime.getRuntime().totalMemory()) && (maxCacheSize/2 > bitmapSize)) {
 					Log.w(DEBUG_TAG,"expanding memory tile cache from " + maxCacheSize + " to " + (maxCacheSize + maxCacheSize/2));
-					maxCacheSize = maxCacheSize + maxCacheSize/2;
+					maxCacheSize = (long)maxCacheSize + maxCacheSize/2;
 				} else {
 					throw new StorageException(StorageException.OOM); // can't expand any more
 				}
